% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server_functions.R
\name{add_tryformats}
\alias{add_tryformats}
\title{Add tryFormats}
\usage{
add_tryformats(castlist, formats, names_to_add)
}
\arguments{
\item{castlist}{A named list of types conversions to perform}

\item{formats}{The formats to add to \code{tryFormats}}

\item{names_to_add}{The names of the elements of \code{castlist}
for which to add a \code{tryFormats} element}
}
\value{
The original \code{castlist} where the specified elements
have a new slot \code{tryFormats} which contains the \code{formats} vector.
If the value of \code{castlist} to modify is \code{NULL}, will not add
\code{tryFormats} slot.
}
\description{
Adds a \code{tryFormats} element to the specified elements of
\code{castlist}.
}
