% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_app.R
\name{allspeciesApp}
\alias{allspeciesApp}
\title{Run standalone allspecies module}
\usage{
allspeciesApp(...)
}
\arguments{
\item{...}{Arguments to pass to shiny::shinyApp besides server and ui}
}
\value{
A working Shiny app
}
\description{
Run standalone allspecies module
}
