% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/camtrapviz-package.R
\docType{package}
\name{camtrapviz-package}
\alias{camtrapviz}
\alias{camtrapviz-package}
\title{camtrapviz: Vizualise Camera Trap Data}
\description{
Compute and visualize common descriptors for camera trap data. A Shiny app allowing to do so interactively and without any code is also implemented.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://lisanicvert.github.io/camtrapviz/}
}

}
\author{
\strong{Maintainer}: Lisa Nicvert \email{lisa.nicvert@univ-lyon1.fr}

}
\keyword{internal}
