% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server_functions.R
\name{cast_columns}
\alias{cast_columns}
\title{Cast columns to expected types}
\usage{
cast_columns(df, col_mapping)
}
\arguments{
\item{df}{A dataframe containing the columns specified in col_mapping (values)}

\item{col_mapping}{A named character vector: names are toe codes and values are
the corresponding column names}
}
\value{
the df with casted columns
}
\description{
Cast columns to expected types
}
\examples{
library(camtraptor)
data(mica)
mapping <- c("col_spp" = "vernacularNames.en",
             "cam_col" = "deploymentID",
             "timestamp_col" = "timestamp")
cast_columns(mica$data$observations, mapping)
}
