% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_ui_functions.R
\name{create_dashboard}
\alias{create_dashboard}
\title{Create a dashboard}
\usage{
create_dashboard(tagList, menu_title, dashboard_title = "Camtrapviz")
}
\arguments{
\item{tagList}{The tagList}

\item{menu_title}{The title to display for the tab menu}

\item{dashboard_title}{Dashboard title}
}
\value{
The tagList wrapped in the dashboard
}
\description{
Wrap a tagList inside a dashboard (used for testing modules)
}
