% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_data_wrangling.R
\name{filter_data}
\alias{filter_data}
\title{Filter camera trap data}
\usage{
filter_data(
  dat,
  spp_filter = NULL,
  spp_col = NULL,
  obs_filter = NULL,
  obs_col = NULL,
  cam_filter = NULL,
  cam_col_rec = NULL,
  cam_col_cam = cam_col_rec,
  daterange = NULL,
  time_col = NULL,
  cameras_as_factor = FALSE
)
}
\arguments{
\item{dat}{The data to filter It can be either a list with one component \verb{$data}
or a \code{datapackage} object (inheriting list). Either way, the data
are in the \verb{$data} slot with two components:
\itemize{
\item \verb{$deployments} (cameras table)
\item \verb{$observations} (records table)
}}

\item{spp_filter}{Species to filter from the data}

\item{spp_col}{Name of the species column}

\item{obs_filter}{Observation types to filter from the data}

\item{obs_col}{Name of the observation column}

\item{cam_filter}{Cameras to filter from the data}

\item{cam_col_rec}{Name of the cameras column in records table (\code{dat$data$observations}).}

\item{cam_col_cam}{Name of the cameras column in cameras table (\code{dat$data$deployments}).
Defaults to the same value as \code{cam_col_rec}.}

\item{daterange}{Date range to filter on for the data. Can be either a Date
or a POSIX.}

\item{time_col}{Name of the time column (must contain dates or POSIX)}

\item{cameras_as_factor}{Transform cameras as factors?}
}
\value{
The filtered data. Species and dates remove data only in \code{dat$data$observations},
but cameras also remove cameras from \code{dat$data$deployments}.
}
\description{
Allows to filter camera trap data observations and
cameras metadata based on species, cameras and dates.
}
\details{
For the \code{spp_filter} and \code{cam_filter} values:
\itemize{
\item If \code{NULL}, data are not filtered on that condition
(also the case for \code{daterange}).
\item If character(0), all values will be filtered out.
}
}
\examples{
data("recordTableSample", package = "camtrapR")
recordTableSample$DateTimeOriginal <- as.POSIXct(recordTableSample$DateTimeOriginal)
data("camtraps", package = "camtrapR")
dat <- list(data = list(observations = recordTableSample,
                        deployments = camtraps))
# Filter out data for species PBE and VTA, camera Station A and keep
# only data from 2009-05-01 to 2009-05-15.
filter_data(dat, 
            spp_col = "Species", 
            spp_filter = c("PBE", "VTA"),
            cam_col_rec = "Station", 
            cam_filter = "StationA",
            daterange = as.Date(c("2009-05-01", "2009-05-15")),
            time_col = "DateTimeOriginal")
}
