% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server_functions.R
\name{find_default_colname}
\alias{find_default_colname}
\title{Find default colname}
\usage{
find_default_colname(
  pattern,
  colnames,
  empty_allowed,
  empty_placeholder = "Not present in data"
)
}
\arguments{
\item{pattern}{a regular expression to match against colnames}

\item{colnames}{a vector of column names}

\item{empty_allowed}{is empty character allowed for this regex?}

\item{empty_placeholder}{What is the empty placeholder?}
}
\value{
All matched columns names (if there was at least a match)
or the empty_placeholder (if empty_allowed and no match) or NULL
(empty not allowed and no match)
}
\description{
Finds the column name to default to among colnames
}
\examples{
find_default_colname("species", 
                     colnames = c("Species", "cameraID", "DateTime"), 
                     empty_allowed = TRUE)
find_default_colname("foo", 
                     colnames = c("Species", "cameraID", "DateTime"), 
                     empty_allowed = FALSE)
find_default_colname("foo", 
                     colnames = c("Species", "cameraID", "DateTime"), 
                     empty_allowed = TRUE)
}
