% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server_functions.R
\name{find_default_colnames}
\alias{find_default_colnames}
\title{Find default column names}
\usage{
find_default_colnames(
  regex_list,
  colnames,
  empty_allowed_list = list(),
  empty_placeholder = "Not present in data"
)
}
\arguments{
\item{regex_list}{a named list with names corresponding to those of
widgets and values are regular expressions used to find those widgets
in colnames}

\item{colnames}{a vector of column names}

\item{empty_allowed_list}{a vector of allowed widgets that can be set
to empty}

\item{empty_placeholder}{the empty placeholder}
}
\value{
A named list containing either (the first) matched colname,
NULL (no match and empty not allowed)
or the empty_placeholder (no match and empty allowed)
}
\description{
Finds the column names to default to among colnames
}
\examples{
library(camtraptor)
data(mica)
colnames <- colnames(mica$data$observations)
regex <- c("^vernacularNames\\\\.en$|species", "station|deployment|camera",
           "timestamp|datetime")
names(regex) <-  c("spp_col", "cam_col", "timestamp_col")
find_default_colnames(regex_list = regex,
                      colnames = colnames)
}
