% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server_functions.R
\name{format_table}
\alias{format_table}
\title{Format table}
\usage{
format_table(df, mapping, cast_type)
}
\arguments{
\item{df}{The dataframe to format}

\item{mapping}{The mapping list for columns in the dataframe.
Names are the column types (common between \code{mapping} and
\code{cast_type}). Values are the corresponding column names
in the dataframe.}

\item{cast_type}{A named list containing the name of the
function to cast between types.
The list's names are the names of the columns column types
(common between \code{mapping} and \code{cast_type}).
Elements of this list can be:
\itemize{
\item a character giving a valid function name to call
\item a list with the first element being the function to call (character)
and additional arguments to the function call (that can be named as
the names of the functions' arguments).
}}
}
\value{
The dataframe df with reordered and casted columns.
}
\description{
Casts columns as indicated in \code{cast_type} and
moves columns indicated in \code{mapping} to the beginning
of the table.
}
\examples{
# Create synthetic dataset
df <- data.frame(camera = c("A", "B", "C"),
                 lat = c("20.12", "20.22", "22.34"),
                 lon = c("33.44", "33.45", "33.42"),
                 setup = c("2022-01-01", "2022-01-01", "2022-01-02"))
mapping <- list(cam_col = "camera",
                lat_col = "lat",
                lon_col = "lon",
                setup_col = "setup")
type <- list(cam_col = "as.character",
             lat_col = "as.numeric",
             lon_col = "as.numeric",
             setup_col = list("as.Date",
                              format = "\%Y-\%m-\%d"))
# Format table
format_table(df, mapping, type)
}
