% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server_functions.R
\name{get_cameras_not_in}
\alias{get_cameras_not_in}
\title{Get cameras not in}
\usage{
get_cameras_not_in(dfrecords, dfcameras, cam_col_records, cam_col_cameras)
}
\arguments{
\item{dfrecords}{records dataframe}

\item{dfcameras}{cameras dataframe}

\item{cam_col_records}{name of the cameras column in the records dataframe}

\item{cam_col_cameras}{name of the cameras column in the cameras dataframe}
}
\value{
A named list with 2 components
not_in_records: cameras from dfcameras that are not in dfrecords
not_in_cameras: cameras from dfrecords that are not in dfcameras
If the cameras are toe same in both files, the components of the list are
character vectors of length zero.
}
\description{
From a records and a cameras dataframe, determine which cameras are in a file
but not in the other
}
\examples{
dfrecords <- data.frame(camID = letters[2:7])
dfcam <- data.frame(cameras = letters[1:5])
get_cameras_not_in(dfrecords = dfrecords, 
                   dfcameras = dfcam,
                   cam_col_records = "camID",
                   cam_col_cameras = "cameras")
}
