% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_data_wrangling.R
\name{get_diversity_indices}
\alias{get_diversity_indices}
\title{Get diversity indices}
\usage{
get_diversity_indices(
  count_df,
  spp_col,
  cam_col,
  count_col = "count",
  prop_col = "prop"
)
}
\arguments{
\item{count_df}{Dataframe summarizing counts per species. Can be
computed using the records dataframe with the \code{get_diversity_table}
function.}

\item{spp_col}{Name of the column containing species names}

\item{cam_col}{Name of the column containing cameras names}

\item{count_col}{Name of the column containing species counts}

\item{prop_col}{Name of the column containing species proportions}
}
\value{
A dataframe with one row per camera summarizing diversity indices.
}
\description{
From a summary table of camera/species, return richness, Shannon
and Simpson diversity indices.
}
\details{
Richness is computed as the number of the different species seen at the
camera.
Shannon index is computed as: -\sum{p_i\times ln(p_i)}. It ranges
between 0 and +\inf, zero indicating the lowest diversity.
Simpson index is computed as: \frac{\sum{n_i\times(n_i - 1)}}{N\times(N-1)}.
It ranges between 0 and 1, 1 indicating the lowest diversity.
}
\examples{
countdf <- data.frame(camera = c("C1", "C1", "C1", "C2",
                                 "C3", "C3", "C3"),
                      species = c("cat", "cow", "rabbit",
                                  "cat", "cat", "cow", "rabbit"),
                      count = c(30, 30, 30, 30, 88, 1, 1),
                      prop = c(1/3, 1/3, 1/3, 1, 88/90, 1/90, 1/90))
get_diversity_indices(countdf, 
                     spp_col = "species", cam_col = "camera")
}
