% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_data_wrangling.R
\name{get_diversity_table}
\alias{get_diversity_table}
\title{Get diversity table}
\usage{
get_diversity_table(
  df,
  cam_col,
  spp_col,
  count_col = NULL,
  keep_all_levels = TRUE
)
}
\arguments{
\item{df}{The dataframe}

\item{cam_col}{Name of the column containing cameras names}

\item{spp_col}{Name of the column containing species names}

\item{count_col}{Name of the column containing species counts (optional)}

\item{keep_all_levels}{Should we keep all camera levels even
those that have no rows in \code{df}? (Checked only if \code{cam_col} is a factor)}
}
\value{
A dataframe with one row per camera-species combination and with the
count of individuals of each species seen at this camera.
}
\description{
Compute a table with various diversity indices from a species
occurrence dataframe.
}
\examples{
df <- data.frame(species = c("zebra", "cat", "cat", "cow", "cow", NA, NA),
                 type = c("animal", "animal", "animal", "animal", "animal", "human", "blank"),
                 camera = c("C1", "C1", "C1", "C1", "C2", "C2", "C3"),
                 count = c(1, 1, 1, 50, 3, 1, NA))
get_diversity_table(df,
                    cam_col = "camera",
                    spp_col = "species",
                    count_col = "count")
}
