% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server_functions.R
\name{get_example_mapping}
\alias{get_example_mapping}
\title{Get example mapping}
\usage{
get_example_mapping(df, col)
}
\arguments{
\item{df}{a dataframe with columns col and widget}

\item{col}{the name of the column to extract example mapping from}
}
\value{
A named character vector, excluding NA values in col
}
\description{
Return the mapping for example datasets in a vector form
}
