% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.R
\name{get_named_list}
\alias{get_named_list}
\title{Get named vector}
\usage{
get_named_list(df, col, widget_values)
}
\arguments{
\item{df}{The dataframe. Must have a column named widget
and a column named like col.}

\item{col}{The column of the dataframe to extract}

\item{widget_values}{The widgets to get the vector for}
}
\value{
A named list containing the values of col,
names are the widget names
}
\description{
Get a vector named with values in column "widget"
from a dataframe
}
