% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server_functions.R
\name{get_regex_vector}
\alias{get_regex_vector}
\title{Get regex vector}
\usage{
get_regex_vector(df, widget_values)
}
\arguments{
\item{df}{The dataframe}

\item{widget_values}{The widgets to get the regex for}
}
\value{
A named vector containing regex, names are the widget names
}
\description{
Get the regex vector from a dataframe
}
