% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_data_wrangling.R
\name{get_species}
\alias{get_species}
\title{Get unique species from a dataframe}
\usage{
get_species(
  df,
  spp_col,
  obs_col = NULL,
  animal_code = "animal",
  return_df = TRUE
)
}
\arguments{
\item{df}{The dataframe}

\item{spp_col}{name of the species column from the dataframe}

\item{obs_col}{name of the observation type column from the dataframe}

\item{animal_code}{value of \code{obs_col} coding for animal observations.}

\item{return_df}{renturn a dataframe? If \code{TRUE}, will return a dataframe
(see below); else will return a character vector of species names.}
}
\value{
If \code{return_df} is \code{TRUE}, returns a dataframe containing
unique species and their characteristics. Else, returns only the character
vector containing to the column \code{spp_col} of this dataframe.
\itemize{
\item a column named like \code{spp_col} containing unique species names.
\item if \code{obs_col} is not \code{NULL}: a column named like \code{obs_col} containing corresponding
observation type.
The rownames of this table are of the form ID_number and are a unique ID for
the species.
}
}
\description{
Get unique species from a dataframe
}
\details{
When \code{obs_col} is provided, the values of the \code{spp_col} column that are
\code{NA} when \code{obs_col} corresponds to a non-animal species (which is the
case in the \code{camtrapDP} standard) are replaced with the value from
\code{obs_col}.
}
\examples{
df <- data.frame(species = c("rabbit", "cat", "cat", NA, NA, 
                             "cameratrapper", "tourist"),
                 type = c("animal", "animal", "animal", "fire", "blank", 
                          "human", "human"))
get_species(df, spp_col = "species", obs_col = "type")
}
