% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_app.R
\name{importApp}
\alias{importApp}
\title{Run standalone import module}
\usage{
importApp(...)
}
\arguments{
\item{...}{Arguments to pass to shiny::shinyApp besides server and ui}
}
\value{
A working Shiny app
}
\description{
Run standalone import module
}
