% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_plot.R
\name{plot_diversity}
\alias{plot_diversity}
\title{Plot species diversity at cameras}
\usage{
plot_diversity(df, div_col, cam_col, cam_vec = NULL, interactive = FALSE)
}
\arguments{
\item{df}{The dataframe with diversity indices per camera}

\item{div_col}{Name of the column containing the diversity index to plot}

\item{cam_col}{Name of the column containing the cameras names}

\item{cam_vec}{A character vector of all cameras that should appear
on the plot (optional)}

\item{interactive}{Make the plot interactive?}
}
\value{
A ggplot object representing diversity indices
as bars (in x) following the different cameras (in y).
}
\description{
Plot species diversity at cameras
}
\examples{
# Create synthetic data ---
df <- data.frame("camera" = c("C1", "C2", "C4", "C3"),
                 "count" = c(1, 4, 2, 22))

plot_diversity(df, 
               div_col = "count", 
               cam_col = "camera")
}
\seealso{
\href{https://lisanicvert.github.io/camtrapviz/articles/plots.html}{\code{vignette("plots", package = "camtrapviz")}}
}
