% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server_functions.R
\name{plot_points}
\alias{plot_points}
\title{Plot points}
\usage{
plot_points(df, camera_col, timestamp_col, spp_col)
}
\arguments{
\item{df}{The dataframe}

\item{camera_col}{Name of the camera column}

\item{timestamp_col}{Name of the timestamp column}

\item{spp_col}{Name of the species column}
}
\value{
A ggplot
}
\description{
Plot occurrences points from a dataframe
}
\examples{
library(camtrapR)
data("recordTableSample")
plot_points(recordTableSample, 
            camera_col = "Station", 
            timestamp_col = "DateTimeOriginal", 
            spp_col = "Species")
}
