% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server_functions.R
\name{plot_species_bars}
\alias{plot_species_bars}
\title{Plot species bars}
\usage{
plot_species_bars(df, spp_col)
}
\arguments{
\item{df}{The dataframe}

\item{spp_col}{Name of the species column}
}
\value{
A ggplot
}
\description{
Plot the barplot of species abundance from a dataframe
}
\examples{
library(camtrapR)
data("recordTableSample")
plot_species_bars(recordTableSample,
                  spp_col = "Species")
}
