% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server_functions.R
\name{read_csv}
\alias{read_csv}
\title{Read csv}
\usage{
read_csv(file_path, column_separator)
}
\arguments{
\item{file_path}{a valid path to a csv file}

\item{column_separator}{column separator character}
}
\value{
A list with 2 elements:
dat = the dataframe read from the file and
sep = the character separator used to read the file
}
\description{
Reads a csv file from a fileInput widget with the column separator specified in a
radioButtons widget.
}
