% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server_functions.R
\name{remove_rows_with_NA}
\alias{remove_rows_with_NA}
\title{Remove rows with NAs}
\usage{
remove_rows_with_NA(df, mapping)
}
\arguments{
\item{df}{The dataframe to clean}

\item{mapping}{The mapping for columns in the dataframe.}
}
\value{
The dataframe where rows in mapping contain no NA values
(except spp_col which is allowed to have NA values if obs_col is in mapping,
when obs_col != 'animal')
}
\description{
Remove rows with NAs
}
\examples{
library(camtraptor)
data(mica)
df <- mica$data$observations
mapping <- c("spp_col" = "vernacularNames.en",
             "cam_col" = "deploymentID",
             "timestamp_col" = "timestamp",
             "obs_col" = "observationType")
# Purposedly add NAs to chack rows are removed
df$vernacularNames.en[9] <- NA
df$timestamp[1] <- NA
remove_rows_with_NA(df, mapping)
}
