% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_data_wrangling.R
\name{reorder_named_values}
\alias{reorder_named_values}
\title{Reorder values}
\usage{
reorder_named_values(vec, names, keep_all_names = TRUE)
}
\arguments{
\item{vec}{A named vector to reorder}

\item{names}{The order of the names}

\item{keep_all_names}{If some values of \code{names} are not in
the names of \code{vec}, should they be kept?}
}
\value{
The reordered vector \code{vec} with only names in \code{names}.
Some values can be \code{NA} if \code{keep_all_names} is \code{TRUE}.
}
\description{
Reorder values
}
