% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_app.R
\name{run_camtrapviz}
\alias{run_camtrapviz}
\title{Run camtrapviz}
\usage{
run_camtrapviz(...)
}
\arguments{
\item{...}{Arguments to pass to shiny::shinyApp besides server and ui}
}
\value{
A working Shiny app
}
\description{
Runs the Shiny app to vizualise camera trap data
}
