% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_ui_functions.R
\name{select_values}
\alias{select_values}
\title{Select values}
\usage{
select_values(prefix, item)
}
\arguments{
\item{prefix}{the prefix to use for radioButtons and select(ize)Input
widgets}

\item{item}{The item name to display in the widgets labels}
}
\value{
A taglist of widgets
}
\description{
Function to create a set of widgets used to select
values based on a column or based o the values of another column
}
