% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_ui_functions.R
\name{separator_widget}
\alias{separator_widget}
\title{Choose separator}
\usage{
separator_widget(prefix)
}
\arguments{
\item{prefix}{The prefix to use for the widget id (will be "prefix_sep")}
}
\value{
A checkboxGroupInput with id "prefix_sep" to choose between comma,
tabulation and semicolon for the file separator
}
\description{
Function to create a widget to choose a column separator
}
