% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_server.R
\name{server}
\alias{server}
\title{Server}
\usage{
server(input, output, session)
}
\arguments{
\item{input}{Shiny input}

\item{output}{Shiny output}

\item{session}{Shiny session}
}
\value{
The server generating functions for Shiny
}
\description{
Server
}
