% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_app.R
\name{summaryApp}
\alias{summaryApp}
\title{Run standalone summary module}
\usage{
summaryApp(...)
}
\arguments{
\item{...}{Arguments to pass to shiny::shinyApp besides server and ui}
}
\value{
A working Shiny app
}
\description{
Run standalone summary module
}
