% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_data_wrangling.R
\name{time_to_circular}
\alias{time_to_circular}
\title{Time to radians}
\usage{
time_to_circular(time, units = c("radians", "hours"), circular = FALSE)
}
\arguments{
\item{time}{The time of the day (must be an object of class
\code{times} from the \code{chron} package)}

\item{units}{The unit to use (subset of values used in in \code{circular::circular})}

\item{circular}{Return an object of class circular?}
}
\value{
the times converted to the desired unit.
If \code{circular} is \code{TRUE}, returns an object of class \code{circular}.
}
\description{
Convert a time of day to circular data.
}
\examples{
time_to_circular(chron::times(c("00:00:00", "12:00:00", "08:33:44")))
}
