% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_data_wrangling.R
\name{vonMises_density}
\alias{vonMises_density}
\title{von Mises density}
\usage{
vonMises_density(
  mod,
  x = circular::circular(seq(0, 2 * pi, by = 0.01), units = "radians"),
  unit = c("hour", "radians")
)
}
\arguments{
\item{mod}{The model}

\item{x}{The vector on which to fit. Must be of type \code{circular}.}

\item{unit}{Unit to use in the return dataframe. Can be either \code{radians}
or \code{hours}.}
}
\value{
A dataframe with 2 columns:
\itemize{
\item \code{density}: value of the density probability function
\item \code{x}: corresponding value of the variable
(using the unit specified in \code{unit}).
}
}
\description{
Get the density of a von Mises mixture model.
}
\examples{
data("recordTableSample", package = "camtrapR")
recordTableSample <- recordTableSample[recordTableSample$Species == "PBE", ]
recordTableSample$Time <- chron::times(recordTableSample$Time)
mod <- fit_vonMises(recordTableSample$Time, k = 3)
dt <- vonMises_density(mod)
# Visual check
hist(as.numeric(recordTableSample$Time)*2*pi, 
     freq = FALSE, breaks = seq(0, 2*pi, by = pi/8))
lines(as.numeric(dt$x), dt$density, type = "l")
}
