% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server_functions.R
\name{add_tryformats}
\alias{add_tryformats}
\title{Add tryFormats to cast list}
\usage{
add_tryformats(castlist, formats, names_to_add)
}
\arguments{
\item{castlist}{A named list of types conversions to perform}

\item{formats}{The formats to add to tryFormats}

\item{names_to_add}{The names of the list for which to add
a tryFormats element}
}
\value{
The original castlist where elements names like in names_to_add
have a new slot tryFormats = formats.
Will not add tryFormats if the casting function is NULL.
}
\description{
This function adds a tryFormats attribute to some elements of castvec.
}
