% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/camtrapviz-package.R
\docType{package}
\name{camtrapviz-package}
\alias{camtrapviz}
\alias{camtrapviz-package}
\title{camtrapviz: Vizualise Camera Trap Data}
\description{
This package implements several functions to vizualise common descriptors and graphics for camera trap data. It also implements a Shiny app allowing to do so interactively and without any code.
}
\author{
\strong{Maintainer}: Lisa Nicvert \email{lisa.nicvert@univ-lyon1.fr}

}
\keyword{internal}
