% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server_functions.R
\name{cast_columns}
\alias{cast_columns}
\title{Cast columns to expected types}
\usage{
cast_columns(df, cast_type)
}
\arguments{
\item{df}{A dataframe containing the columns specified in col_mapping (values)}

\item{cast_type}{A named vector with the type conversion to perform.
Must be a valid function name to call.
Names are the names of the columns to cast in df.}
}
\value{
the df with casted columns
}
\description{
Cast columns to expected types
}
\examples{
df <- data.frame(num = 1:10,
                 char = letters[1:10])
cast <- c(num = "as.character",
          char = "as.factor")
dfcast <- cast_columns(df, cast)
}
