% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server_functions.R
\name{cast_columns}
\alias{cast_columns}
\title{Cast columns to expected types}
\usage{
cast_columns(df, cast_type)
}
\arguments{
\item{df}{A dataframe containing the columns specified in col_mapping (values)}

\item{cast_type}{A named list with the type conversion to perform.
Names are the names of the columns to cast in df.
Elements of this list can either be characters giving
a valid function name to call, or a list with the first element as
the function to call and additional arguments (that can be named)
being the arguments to pass to the function.}
}
\value{
the df with casted columns
}
\description{
Cast columns to expected types
}
\examples{
df <- data.frame(num = 1:10,
                 char = letters[1:10])
cast <- list(num = "as.character",
             char = "as.factor")
dfcast <- cast_columns(df, cast)
}
