% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_format_data.R
\name{clean_data}
\alias{clean_data}
\title{Clean data}
\usage{
clean_data(
  dat,
  cam_type,
  rec_type,
  only_shared_cameras = FALSE,
  cam_col_records,
  cam_col_cameras = NULL,
  split = FALSE,
  add_rowid = FALSE
)
}
\arguments{
\item{dat}{The data to clean. It can be either a list with one component \verb{$data}
or a \code{datapackage} object (inheriting list). Either way, the data
are in the \verb{$data} slot with two components:
\itemize{
\item \verb{$deployments} (cameras table)
\item \verb{$observations} (records table)
}}

\item{cam_type}{A named list containing the name of the
function to cast between types for the cameras table.
The list's names are the names of the columns to cast
in \code{dat$data$deployments}.
For details on the content of this list, see the documentation of
the \code{cast_columns} function.}

\item{rec_type}{A named list containing the name of the
function to cast between types for the records table.
The list's names are the names of the columns to cast
in \code{dat$data$observations}.
For details on the content of this list, see the documentation of
the \code{cast_columns} function.}

\item{only_shared_cameras}{Logical; restrict final data to shared cameras
that are in \code{dat$data$deployments} and in \code{dat$data$observations}?}

\item{cam_col_records}{Name of the column with cameras names in
records (needed only of \code{only_shared_cameras} is \code{TRUE})}

\item{cam_col_cameras}{Name of the column with cameras names in
cameras (needed only of \code{only_shared_cameras} is \code{TRUE}).
If \code{NULL} will be assumed to be the same as \code{cam_col_records}.}

\item{split}{Logical; should the camera data be extracted from the
records table by splitting the data?}

\item{add_rowid}{Should row IDs be added to the observations df?
If yes, row names in the form of "ID_xx" are added to the the dataframe.}
}
\value{
An object of the same type as the original input,
but where \code{dat$data$deployments} and \code{dat$data$observations} have been
cleaned as described above.
}
\description{
Cleans data by:
\itemize{
\item splitting records data in records and cameras (if needed)
\item formatting cameras and records tables: casting columns and
moving the selected columns to the beginning
\item if \code{only_shared_cameras} is \code{TRUE}: selecting the subset of
cameras present in both records and cameras tables
}
}
\examples{
# Create synthetic dataset
records <- data.frame(foo = 1:6,
                      species = c("pigeon", "mouse", "pigeon", "mouse", "mouse", "pigeon"),
                      date = c("2022-01-01", "2022-03-01", 
                               "2022-01-02", "2022-01-12", "2022-01-22",
                               "2022-01-03"),
                      time = c("10:22:01", "22:12:01",
                               "11:54:33", "07:14:38", "18:01:34", 
                               "12:11:34"),
                      camera = c("A", "A", "B", "B", "B", "C"))
cameras <- data.frame(camera = c("A", "B", "C"),
                      lat = c("20.12", "20.22", "22.34"),
                      lon = c("33.44", "33.45", "33.42"))
dat <- list(data = list(observations = records,
                        deployments = cameras))
rec_type <- list(species = "as.character",
                 date = list("as.Date",
                                 format = "\%Y-\%m-\%d"),
                 time = "times",
                 camera = "as.character")
cam_type <- list(camera = "as.character",
                 lat = "as.numeric",
                 lon = "as.numeric")
# Clean data converts columns to the appropriate types 
# and reorders columns
clean_data(dat,
           rec_type = rec_type,
           cam_type = cam_type)
}
