% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server_functions.R
\name{clean_data}
\alias{clean_data}
\title{Clean data}
\usage{
clean_data(
  dat,
  mapping_cameras,
  cam_type,
  mapping_records,
  rec_type,
  split = FALSE
)
}
\arguments{
\item{dat}{The data ti clean}

\item{mapping_cameras}{The mapping for columns in the cameras dataframe.}

\item{cam_type}{A named vector with the type conversion to perform.
Must be a valid function name to call.
Names are the names of the columns to cast in cameras df.}

\item{mapping_records}{The mapping for columns in the records dataframe.}

\item{rec_type}{A named vector with the type conversion to perform.
Must be a valid function name to call.
Names are the names of the columns to cast in records df.}

\item{split}{Should the camera data be splitted from the records table?}
}
\value{
The cleaned dataset
}
\description{
Cleans raw data by:
\itemize{
\item splitting data if needed
\item formatting cameras and records tables
\item selecting the subset of cameras present in both datasets
}
}
