% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_format_data.R
\name{filter_cameras_in_both_tables}
\alias{filter_cameras_in_both_tables}
\title{Filter cameras in both tables}
\usage{
filter_cameras_in_both_tables(
  dfrec,
  dfcam,
  cam_col_dfrec,
  cam_col_dfcam = NULL
)
}
\arguments{
\item{dfrec}{Records dataframe}

\item{dfcam}{Cameras dataframe}

\item{cam_col_dfrec}{Name of the column with cameras names in \code{dfrec}}

\item{cam_col_dfcam}{Name of the column with cameras names in \code{dfcam}
If \code{NULL} will be assumed to be the same as \code{cam_col_dfrec}.}
}
\value{
A list of two dataframes with filtered values:
\itemize{
\item \verb{$records} is the records dataframe
\item \verb{$cameras} is the cameras dataframe
}
}
\description{
Filter data to keep only rows where cameras are in both tables
}
\examples{
dfrec <- data.frame(species = c("pigeon", "mouse", "pigeon", "mouse"),
                      stamp = Sys.time() + seq(60, length.out = 4, by = 60),
                      camera = c("A", "B", "C", "E"))
dfcam <- data.frame(camera = c("A", "B", "C", "D"), 
                      lat = c(20.12, 20.22, 22.34, 21.35),
                      lon = c(33.44, 33.45, 33.42, 33.53))
filter_cameras_in_both_tables(dfrec, dfcam, 
                              cam_col_dfrec = "camera")
}
