% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server_functions.R
\name{find_default_colname}
\alias{find_default_colname}
\title{Find default colname}
\usage{
find_default_colname(
  pattern,
  colnames,
  empty_allowed,
  empty_placeholder = "Not present in data"
)
}
\arguments{
\item{pattern}{a regular expression to look for in \code{colnames}}

\item{colnames}{a vector of column names}

\item{empty_allowed}{logical; is it allowed to return
the character in \code{empty_placeholder} if there is no match?}

\item{empty_placeholder}{a character to use as placeholder
if there is no match for \code{pattern} in \code{columns}.}
}
\value{
\itemize{
\item if there was at least a match: all matched columns names
\item if there was no match:
\itemize{
\item if \code{empty_allowed}, returns the empty placeholder
\item else, returns \code{NULL}
}
}
}
\description{
Finds a column name matching the given pattern among a vector
of column names given in argument.
}
\examples{
# There is a match
camtrapviz:::find_default_colname("species",
                                  colnames = c("Species", "cameraID", "DateTime"),
                                  empty_allowed = TRUE)
}
