% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server_functions.R
\name{find_default_colnames}
\alias{find_default_colnames}
\title{Find default column names}
\usage{
find_default_colnames(
  colnames,
  empty_allowed_list = list(),
  empty_placeholder = "Not present in data"
)
}
\arguments{
\item{colnames}{a vector of column names}

\item{empty_allowed_list}{a vector of allowed widgets that can be set
to empty}

\item{empty_placeholder}{the empty placeholder}
}
\value{
A named list containing either (the first) matched colname,
NULL (no match and empty not allowed)
or the empty_placeholder(no match and empty allowed)
}
\description{
Finds the column names to default to among colnames
}
\details{
currently implemented for the followind widgets:
spp_col, cam_col, date_col, time_col, timestamp_col,
count_col, lat_col, long_col
}
\examples{
library(camtraptor)
data(mica)
colnames <- colnames(mica$data$observations)
find_default_colnames(colnames)
}
