% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_data_wrangling.R
\name{fit_vonMises}
\alias{fit_vonMises}
\title{Fit a von Mises distribution.}
\usage{
fit_vonMises(time, k)
}
\arguments{
\item{time}{The time of the day (must be an object of class
\code{times} from the \code{chron} package)}

\item{k}{Number of mixxture components (number of modes
in the final ditribution)}
}
\value{
A mixture moden of von Mises distributions of class
\code{movMF}.
}
\description{
Fit a von Mises distribution.
}
\examples{
data("recordTableSample", package = "camtrapR")
recordTableSample <- recordTableSample[recordTableSample$Species == "PBE", ]
recordTableSample$Time <- chron::times(recordTableSample$Time)
fit_vonMises(recordTableSample$Time, k = 3)
}
