% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server_functions.R
\name{format_table}
\alias{format_table}
\title{Cleans a dataframe}
\usage{
format_table(df, mapping)
}
\arguments{
\item{df}{The dataframe to clean}

\item{mapping}{The mapping for columns in the dataframe.}
}
\value{
The dataframe df with cleaned columns.
}
\description{
Moves columns indicated in mapping to the beginning,
and casts those columns.
}
\examples{
library(camtraptor)
data(mica)
mapping <- c("col_spp" = "vernacularNames.en",
             "cam_col" = "deploymentID",
             "timestamp_col" = "timestamp")
format_table(mica$data$observations, mapping)
}
