% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_data_wrangling.R
\name{get_diversity_indices}
\alias{get_diversity_indices}
\title{Get diversity indices}
\usage{
get_diversity_indices(
  count_df,
  spp_col,
  cam_col,
  count_col = "individuals",
  prop_col = "individuals_prop"
)
}
\arguments{
\item{count_df}{Dataframe summarizing counts per species. Can be
computed using the records dataframe with the \code{summarize_species}
function.}

\item{spp_col}{Name of the column containing species names}

\item{cam_col}{Name of the column containing cameras names}

\item{count_col}{Name of the column containing species counts}

\item{prop_col}{Name of the column containing species proportions}
}
\value{
A dataframe with one row per camera summarizing diversity indices.
}
\description{
From a summary table of camera/species, return richness, Shannon
and Simpson diversity indices.
}
\details{
Richness is computed as the number of the different species seen at the
camera.

Shannon index is computed as:
\deqn{-\sum p_i ln(p_i)}
Where p_i represents the proportion of the abundance of species i at
a camera over the the total number of individuals of all species
seen at this camera.
It ranges between 0 and +infinity, zero indicating the
lowest diversity.

Simpson index is computed as:
\deqn{(\sum n_i (n_i - 1))/(N (N-1))}
Where n_i represents the abundance of species i at
a camera over the total number of individuals of all species
seen at this camera.
It ranges between 0 and 1, one indicating the lowest diversity.
}
\examples{
countdf <- data.frame(camera = c("C1", "C1", "C1", "C2",
                                 "C3", "C3", "C3"),
                      species = c("cat", "cow", "rabbit",
                                  "cat", "cat", "cow", "rabbit"),
                      individuals = c(30, 30, 30, 30, 88, 1, 1),
                      individuals_prop = c(1/3, 1/3, 1/3, 1, 88/90, 1/90, 1/90))
get_diversity_indices(countdf, 
                      spp_col = "species", cam_col = "camera")
}
