% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_plot.R
\name{plot_map}
\alias{plot_map}
\title{Plot map}
\usage{
plot_map(
  df,
  lat_col,
  lon_col,
  crs = 4326,
  cam_col,
  color = "black",
  highlight_color = "red",
  circle_radii = 10,
  rescale = FALSE
)
}
\arguments{
\item{df}{A dataframe containing cameras information}

\item{lat_col}{Name of the latitude (or the projected y-coordinate)
column}

\item{lon_col}{Name of the longitude (or the projected y-coordinate)
column}

\item{crs}{EPSG code for the coordinate reference system (CRS)
Defaults to EPSG:4326, which is the code for WGS84 standard.}

\item{cam_col}{Name of the camera name column}

\item{color}{color for the points}

\item{highlight_color}{color to use when hovering over a point}

\item{circle_radii}{A named vector of radii tu use for the cirles. Names
correspond to camera names.}

\item{rescale}{rescale circles? If \code{TRUE}, radii will be linearly resized
so that the maximum corresponds to 300m, and radii smaller than
10 will be set to 10m.}
}
\value{
a \code{leaflet} map representing cameras as points.
If the CRS of the input data is different from EPSG:4326 (WGS84),
data are re-projected using WGS84.
When hovering over a camera, it becomes red and its name is shown.
When clicking on a camera, a popup displaying the camera name appears.
}
\description{
Plot a leaflet map representing cameras' coordinates as points.
}
\examples{
data(camtraps, package = "camtrapR")
plot_map(camtraps,
         lat_col = "utm_y", 
         lon_col = "utm_x",
         crs = 32650, # Here we use the EPSG code for UTM zone 50N
         cam_col = "Station")
}
