% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server_functions.R
\name{plot_map}
\alias{plot_map}
\title{Plot map}
\usage{
plot_map(df, lat_col, lon_col, crs = NULL, cam_col, color = "black")
}
\arguments{
\item{df}{A dataframe with latitude and longitude columns}

\item{lat_col}{Name of the latitude column}

\item{lon_col}{Name of the longitude column}

\item{crs}{EPSG code for the coordinate reference system}

\item{cam_col}{Name of the camera name column}

\item{color}{color for the points}
}
\value{
a leaflet map
}
\description{
Plot a leaflet map of cameras
}
