% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_points}
\alias{plot_points}
\title{Plot points}
\usage{
plot_points(
  df,
  camera_col,
  spp_col,
  timestamp_col,
  date_col = NULL,
  time_col = NULL,
  interactive = TRUE
)
}
\arguments{
\item{df}{The dataframe}

\item{camera_col}{Name of the camera column}

\item{spp_col}{Name of the species column}

\item{timestamp_col}{Name of the timestamp column. If it
is a datetime, it must be of class POSIXct. It can be \code{NULL}
if \code{date_col} and \code{time_col} are provided.}

\item{date_col}{Name of the date column. It is assumed
to be of class \code{Date} (else, results are not guaranteed).
It can be \code{NULL} if \code{timestamp_col} is provided.}

\item{time_col}{Name of the time column. It is assumed
to be of class \code{times} (else, results are not guaranteed).
It can be \code{NULL} if \code{timestamp_col} is provided.}

\item{interactive}{Logical; make the plot interactive with \code{ggiraph}?}
}
\value{
A \code{ggplot} object representing time on the x-axis
and cameras on the y-axis. Colors of the points correspond to
different species.
}
\description{
Plot species occurrences at cameras as points
}
\details{
If \code{date_col} and \code{time_col} are provided along with
\code{timestamp_col}, they will be ignored.
}
\examples{
data("recordTableSample", package = "camtrapR")
recordTableSample$DateTimeOriginal <- as.POSIXct(recordTableSample$DateTimeOriginal)
plot_points(recordTableSample, 
            camera_col = "Station", 
            timestamp_col = "DateTimeOriginal", 
            spp_col = "Species")
}
