% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server_functions.R
\name{plot_points}
\alias{plot_points}
\title{Plot points}
\usage{
plot_points(
  df,
  camera_col,
  spp_col,
  timestamp_col,
  date_col = NA,
  time_col = NA,
  interactive = TRUE
)
}
\arguments{
\item{df}{The dataframe}

\item{camera_col}{Name of the camera column}

\item{spp_col}{Name of the species column}

\item{timestamp_col}{Name of the timestamp column
(can be null if date_col and time_col are provided)}

\item{date_col}{Name of the date column (can be NA if timestamp_col is provided)}

\item{time_col}{Name of the time column (can be NA if timestamp_col is provided)}

\item{interactive}{Make the plot interactive?}
}
\value{
A ggplot
}
\description{
Plot occurrences points from a dataframe
}
\details{
If date_col and time_col are provided along timestamp_col,
they will be ignored.
}
\examples{
library(camtrapR)
data("recordTableSample")
plot_points(recordTableSample, 
            camera_col = "Station", 
            timestamp_col = "DateTimeOriginal", 
            spp_col = "Species")
}
