% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server_functions.R
\name{plot_species_bars}
\alias{plot_species_bars}
\title{Plot species bars}
\usage{
plot_species_bars(
  df,
  spp_col,
  count_col = NULL,
  obs_col = NULL,
  interactive = TRUE
)
}
\arguments{
\item{df}{The dataframe}

\item{spp_col}{Name of the species column}

\item{count_col}{Name of the count column}

\item{obs_col}{Name of the observation type column}

\item{interactive}{Make the plot interactive?}
}
\value{
A ggplot
}
\description{
Plot the barplot of species abundance from a dataframe
}
\examples{
library(camtrapR)
data("recordTableSample")
plot_species_bars(recordTableSample,
                  spp_col = "Species")
}
