% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_plot.R
\name{plot_species_bars}
\alias{plot_species_bars}
\title{Plot species bars}
\usage{
plot_species_bars(
  df,
  spp_col,
  count_col = NULL,
  obs_col = NULL,
  interactive = TRUE,
  NA_count_placeholder = NA
)
}
\arguments{
\item{df}{The dataframe}

\item{spp_col}{Name of the species column}

\item{count_col}{Name of the count column (optional). If missing,
it will be assumed to be 1 for all observations.}

\item{obs_col}{Name of the observation type column (optional).
If it is present, the function will plot only the observations
for which \code{obs_col} is "animal".}

\item{interactive}{Logical; make the plot interactive with \code{ggiraph}?}

\item{NA_count_placeholder}{Value with which to replace NAs present
in the column containing counts. If not specified, NA is the default
and species which have NA in counts will have a NA count.}
}
\value{
A \code{ggplot} object representing horizontal bars of species
count. The x-axis is the observed number of individuals and the y-axis
are the different species.
}
\description{
Plot the barplot of species abundance
}
\examples{
data("recordTableSample", package = "camtrapR")
plot_species_bars(recordTableSample,
                  spp_col = "Species")
}
