% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server_functions.R
\name{prepare_cameras}
\alias{prepare_cameras}
\title{Prepare cameras}
\usage{
prepare_cameras(dat, mapping_cameras, split = FALSE)
}
\arguments{
\item{dat}{The data, a list with at least one component
$data
$deployments
$observations}

\item{mapping_cameras}{The mapping for camera columns}

\item{split}{should camera data be extracted from the records?}
}
\value{
The dataset with "pre-cleaned" camera data, i.e.
if split = TRUE, dat$data$deployments is filled with data extracted from the recorde
the records are unique across the mapping column
}
\description{
Prepare the camera data for cleaning
}
