% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_format_data.R
\name{prepare_cameras}
\alias{prepare_cameras}
\title{Prepare cameras}
\usage{
prepare_cameras(dat, cam_columns, split = FALSE)
}
\arguments{
\item{dat}{The data. It can be either a list with one component \verb{$data}
or a \code{datapackage} object (inheriting list). Either way, the data
are in the \verb{$data} slot with two components:
\itemize{
\item \verb{$deployments} (cameras table)
\item \verb{$observations} (records table)
}}

\item{cam_columns}{A list of columns names that indicate
relevant colums to consider for the deployments data.
These columns are moved to the beginning of the cameras table
(\code{dat$data$deployments}).
If \code{split} is \code{TRUE}, these names are also extracted from
\code{dat$data$observations} to create the cameras table.}

\item{split}{should camera data be extracted from the records?
If yes, a cameras table will be created and replace the value
of \code{dat$data$deployments}.}
}
\value{
The dataset with "pre-cleaned" cameras data:
\itemize{
\item the cameras table is filtered to keep unique rows across
the columns indicated in \code{cam_columns}. This allows to filter
for instance between duplicated cameras names.
\item if \code{split} is \code{TRUE}, \code{dat$data$deployments} is replaced with
data extracted from \code{dat$data$observations}
(only the columns indicated in \code{cam_columns}).
}
}
\description{
Prepare the camera data for cleaning
}
