% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server_functions.R
\name{print_check_cameras}
\alias{print_check_cameras}
\title{Print check cameras}
\usage{
print_check_cameras(cameras, type = c("not_in_records", "not_in_cameras"))
}
\arguments{
\item{cameras}{character vector of camera names}

\item{type}{type of message to print: not_in_records or not_in_cameras}
}
\value{
A message describing which cameras are missing where,
the empty string of no cameras are missing
}
\description{
Prints a message to describe cameras status
}
