% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server_functions.R
\name{read_data}
\alias{read_data}
\title{Read data}
\usage{
read_data(
  records_path,
  sep_records,
  cameras_path = NULL,
  sep_cameras = NULL,
  NA_strings = c("NA", "")
)
}
\arguments{
\item{records_path}{A valid file path for records.}

\item{sep_records}{separator used for the records.}

\item{cameras_path}{A valid file path for cameras (defaults to NULL).}

\item{sep_cameras}{separator used for the cameras (defaults to NULL).}

\item{NA_strings}{Vector of characters that should be considered NAs
after import}
}
\value{
A list with 1 components or a camtrapDP list.
If records_path is a json file, returns a captrapDP list.
If records_path is a csv file, returns a list with 1 component:
$data: a list with 2 components:
$observations (records)
$deployments (cameras: if no camera file, is NULL)
}
\description{
Reads data from a file path (either csv of json file), and
optionnally another csv file with camera data.
}
