% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server_functions.R
\name{summarize_cameras}
\alias{summarize_cameras}
\title{Summarize cameras}
\usage{
summarize_cameras(df, cam_col, time_col)
}
\arguments{
\item{df}{the dataframe}

\item{cam_col}{name of the column containing camera IDs}

\item{time_col}{name of the column containing stamps#'}
}
\value{
A summarized dataframe with one row per camera:
cam_col, start, end
}
\description{
Summarize camera trap data (records) by creating a start
and an end date.
}
