% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server_functions.R
\name{summarize_cameras}
\alias{summarize_cameras}
\title{Summarize cameras}
\usage{
summarize_cameras(df, cam_col, timestamp_col, date_col = NULL, time_col = NULL)
}
\arguments{
\item{df}{the dataframe}

\item{cam_col}{name of the column containing camera IDs}

\item{timestamp_col}{name of the column containing timestamps}

\item{date_col}{name of the column containing date}

\item{time_col}{name of the column containing hour}
}
\value{
A summarized dataframe with one row per camera:
cam_col (name of the camera column), setup, retrieval
}
\description{
Summarize camera trap data (records) by creating a start
and an retrieval date.
}
