% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_data_wrangling.R
\name{summarize_species}
\alias{summarize_species}
\title{Summarize species}
\usage{
summarize_species(
  df,
  species_col,
  cam_col = NULL,
  obs_col = NULL,
  count_col = NULL,
  ncam = NULL,
  NA_count_placeholder = NA
)
}
\arguments{
\item{df}{the observation dataframe to summarize}

\item{species_col}{Name of the species column}

\item{cam_col}{Name of the column containing camera codes (optional)}

\item{obs_col}{Name of the observation type column (optional)}

\item{count_col}{Name of the column containing species count (optional)}

\item{ncam}{Number of cameras to take into account when computing
the proportion of cameras the species was ween on. If \code{NULL},
defaults to the number of cameras present in the \code{df}.}

\item{NA_count_placeholder}{Value with which to replace NAs present
in the column containing counts. Defaults to NA (i.e. values are not
replaced)}
}
\value{
A table summarizing species information with the following columns:
\itemize{
\item Species (named like \code{species_col}): species identity
(same as the \code{species_col} input column)
\item Observation type (present only if \code{obs_col} is not \code{NULL} and named like
\code{obs_col}): observation type (same as the \code{obs_col} input column)
\item \code{n_sightings}: number of rows where the species was photographed.
\item \code{n_individuals}: count of individuals observed on all
pictures (using the input \code{count_col} column).
If \code{count_col} is \code{NULL}, it contains the same values as \code{n_sightings}.
If there are NAs in the input  \code{count_col}, they will propagate in \code{n_individuals}
(unless a value is specified in \code{NA_count_placeholder}).
\item \code{n_cameras} (present only if \code{cam_col} is not \code{NULL}) : the number of cameras the species was seen on.
\item \code{prop_cam} (present only if \code{cam_col} is not \code{NULL}): the proportion of cameras the species was seen on.
If \code{ncam} is provided, then it uses \code{ncam} as the total number of cameras.
}
}
\description{
Summarize species information from a data table
}
\examples{
df <- data.frame(species = c("zebra", "cat", "cat", "cow", NA, NA),
                 type = c("animal", "animal", "animal", "animal", "human", "blank"),
                 camera = c("C1", "C1", "C2", "C3", "C3", "C4"),
                 count = c(1, 1, 3, 50, 1, NA))
res <- summarize_species(df, 
                         species_col = "species", cam_col = "camera",
                         obs_col = "type",
                         count_col = "count",
                         NA_count_placeholder = 1)
}
